proj1/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-admin-generator==2.6.0
django-crispy-forms==2.3
django-extensions==3.2.3
django-model-utils==5.0.0
django-utils2==3.0.2
pyparsing==3.2.0
python-utils==3.9.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
